+++ examples/main.rs
@@ -5,7 +5,7 @@ use fonterator::FontChain;
 use footile::{FillRule, Plotter, Raster, Rgba8};
 
 fn main() {
-    // This only succeeds if collection consists of one font
+    // Load the default FontChain (font and fallbacks).
     let font = FontChain::default();
 
     // Init rendering
@@ -15,8 +15,8 @@ fn main() {
     // Render the text
     let path = font.render(
         "Héllö,\nWørłd!‽i\n野ウサギ\nWW野WWウ\nサWWギWW", /*text*/
-        (0.0, 0.0),       /*position*/
-        (256.0, 256.0),   /*size*/
+        (0.0, 0.0),                                                     /*position*/
+        (256.0, 256.0),                                                 /*size*/
     );
     r.over(
         p.fill(path, FillRule::NonZero),