+++ changelog.md
@@ -1,3 +1,7 @@
+## 0.3.0
+* Added `FontChain` for default and fallback fonts.  `FontChain` is an abstraction over the old `Font` and `FontCollection` structs, which are no longer part of the public API.
+* Added proper support for multi-lingual monospace with `FontChain::multilingual_mono()`.
+
 ## 0.2.1
 * Fix README.